Warnings, begone!

+++ ShortcutRecorder/Framework/Source/SRCommon.m
@@ -50,10 +50,10 @@ NSString * SRStringForKeyCode( NSInteger keyCode )
 NSString * SRStringForCarbonModifierFlags( NSUInteger flags )
 {
     NSString *modifierFlagsString = [NSString stringWithFormat:@"%@%@%@%@", 
-		( flags & controlKey ? [NSString stringWithFormat:@"%C", KeyboardControlGlyph] : @"" ),
-		( flags & optionKey ? [NSString stringWithFormat:@"%C", KeyboardOptionGlyph] : @"" ),
-		( flags & shiftKey ? [NSString stringWithFormat:@"%C", KeyboardShiftGlyph] : @"" ),
-		( flags & cmdKey ? [NSString stringWithFormat:@"%C", KeyboardCommandGlyph] : @"" )];
+		( flags & controlKey ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardControlGlyph] : @"" ),
+		( flags & optionKey ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardOptionGlyph] : @"" ),
+		( flags & shiftKey ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardShiftGlyph] : @"" ),
+		( flags & cmdKey ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardCommandGlyph] : @"" )];
 	return modifierFlagsString;
 }
 
@@ -73,10 +73,10 @@ NSString * SRStringForCarbonModifierFlagsAndKeyCode( NSUInteger flags, NSInteger
 NSString * SRStringForCocoaModifierFlags( NSUInteger flags )
 {
     NSString *modifierFlagsString = [NSString stringWithFormat:@"%@%@%@%@", 
-		( flags & NSControlKeyMask ? [NSString stringWithFormat:@"%C", KeyboardControlGlyph] : @"" ),
-		( flags & NSAlternateKeyMask ? [NSString stringWithFormat:@"%C", KeyboardOptionGlyph] : @"" ),
-		( flags & NSShiftKeyMask ? [NSString stringWithFormat:@"%C", KeyboardShiftGlyph] : @"" ),
-		( flags & NSCommandKeyMask ? [NSString stringWithFormat:@"%C", KeyboardCommandGlyph] : @"" )];
+		( flags & NSControlKeyMask ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardControlGlyph] : @"" ),
+		( flags & NSAlternateKeyMask ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardOptionGlyph] : @"" ),
+		( flags & NSShiftKeyMask ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardShiftGlyph] : @"" ),
+		( flags & NSCommandKeyMask ? [NSString stringWithFormat:@"%C", (unsigned short)KeyboardCommandGlyph] : @"" )];
 	
 	return modifierFlagsString;
 }
